# public/openapi.yaml
openapi: 3.1.0
info:
  title: tevaxia.lu API
  version: 1.1.0
  description: >-
    API de calculs indicatifs. Les hypothèses internes ne constituent ni une expertise certifiée ni une
    décision de crédit. Authentification selon chaque endpoint, par X-API-Key ou Authorization: Bearer. Les
    endpoints authentifiés appliquent des quotas et peuvent retourner 429 avec Retry-After.
  contact:
    name: Support tevaxia.lu
    email: contact@tevaxia.lu
    url: https://www.tevaxia.lu
  license:
    name: Proprietary
servers:
  - url: https://tevaxia.lu
    description: Production
security:
  - ApiKeyHeader: []
  - BearerAuth: []
...
paths:
  /api/v1/estimation:
    options:
      summary: Pré-vol CORS
      responses:
        '204':
          description: CORS OK
    post:
      tags:
        - Estimation
      summary: Estimation indicative d’un appartement en pleine propriété
      description: >-
        Référence communale agrégée, ajustements internes non validés statistiquement et fourchette
        conventionnelle. Pas de ventes individuelles, pas de prix par quartier. Les paramètres facultatifs
        omis utilisent les valeurs par défaut du schéma ; aucune coercition de type ni aucun champ inconnu
        n’est accepté.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstimationInput'
            examples:
              luxembourg:
                summary: Appartement existant 85 m² — hypothèses explicites
                value:
                  commune: Luxembourg
                  surface: 85
                  nbChambres: 2
                  etage: adjEtage2e3eRef
                  etat: adjEtatBonRef
                  exterieur: adjExtBalconRef
                  parking: false
                  classeEnergie: D
                  typeBien: appartement
                  estNeuf: false
      responses:
        '200':
          description: Estimation calculée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimationResponse'
        '400':
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Non authentifié
        '404':
          description: Commune ambiguë ou prix indisponible pour le segment
        '429':
          description: Quota dépassé
  /api/v1/estimation/batch:
    options:
      summary: Pré-vol CORS
      responses:
        '204':
          description: CORS OK
    post:
      tags:
        - Estimation
      summary: Estimation en lot (jusqu'à 1000 biens)
      description: >-
        Même validation, valeurs par défaut et méthode que l’endpoint unitaire. Réponse 200 pour un lot
        traité, même si des éléments échouent ; consulter success et error pour chaque index. Aucun résultat
        réglementaire.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
              properties:
                items:
                  type: array
                  maxItems: 1000
                  items:
                    $ref: '#/components/schemas/EstimationInput'
                  minItems: 1
      responses:
        '200':
          description: Lot traité (succès partiels possibles dans results[].error)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  succeeded:
                    type: integer
                  failed:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        index:
                          type: integer
                        success:
                          type: boolean
                        data:
                          type: object
                        error:
                          type: string
        '400':
          description: Corps invalide ou lot vide
        '413':
          description: Lot trop volumineux (> 1000 items)
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    EstimationInput:
      type: object
      required:
        - commune
        - surface
      properties:
        commune:
          type: string
          description: Nom de la commune luxembourgeoise (ex. "Luxembourg")
          minLength: 1
          maxLength: 200
        quartier:
          type: string
          description: Information facultative, sans ajustement par quartier dans le jeu de données actuel
          maxLength: 200
        surface:
          type: number
          maximum: 10000
          description: Surface habitable en m²
          exclusiveMinimum: 0
        nbChambres:
          type: integer
          minimum: 0
          maximum: 20
          default: 0
        etage:
          type: string
          enum:
            - adjEtageSousSol
            - adjEtageRDC
            - adjEtage1er
            - adjEtage2e3eRef
            - adjEtage4e5e
            - adjEtageDernier
            - adjEtageAttique
          default: adjEtage2e3eRef
        etat:
          type: string
          enum:
            - adjEtatNeuf
            - adjEtatRenove
            - adjEtatBonRef
            - adjEtatCorrect
            - adjEtatARenover
            - adjEtatGrosTravaux
          default: adjEtatBonRef
        exterieur:
          type: string
          enum:
            - adjExtPasExterieur
            - adjExtBalconRef
            - adjExtGrandBalcon
            - adjExtTerrasse
            - adjExtJardin
            - adjExtTerrasseJardin
          default: adjExtBalconRef
        parking:
          type: boolean
          default: false
        classeEnergie:
          type: string
          enum:
            - A
            - B
            - C
            - D
            - E
            - F
            - G
          default: D
          description: >-
            Hypothèse de classe utilisée par ce modèle ; A+, H, I et inconnue ne disposent pas d’ajustement
            dans cet endpoint.
        typeBien:
          type: string
          enum:
            - appartement
          default: appartement
        estNeuf:
          type: boolean
          default: false
      additionalProperties: false
    EstimationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            estimationBasse:
              type: number
            estimationCentrale:
              type: number
            estimationHaute:
              type: number
            prixM2Ajuste:
              type: number
            confiance:
              type: string
              enum:
                - faible
              description: Limitation du modèle, pas de probabilité statistique
            ajustements:
              type: array
              items:
                type: object
                properties:
                  labelKey:
                    type: string
                  pct:
                    type: number
            prixM2Base:
              type: number
            totalAjustements:
              type: number
            sourceBase:
              type: string
            confianceNote:
              type: string
        meta:
          type: object
          properties:
            api_key_name:
              type: string
            tier:
              type: string
              enum:
                - free
                - pro
                - enterprise
                - sandbox
            method:
              type: string
            limitations:
              type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string